# app/Http/Requests/CategoryRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     schema="CategoryRequest",
 *     title="Category Request",
 *     description="Category request body data",
 *     required={"name"},
 *     @OA\Property(property="name", type="string", example="Food"),
 *     @OA\Property(property="description", type="string", example="Events related to food and dining.")
 * )
 */
class CategoryRequest extends FormRequest

{


    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules()
    {
        return [
            'name' => 'required|string|max:255',
            'description' => 'nullable|string|max:1000',
        ];
    }
}
/**
 * @OA\Schema(
 *     schema="CategoryRequest",
 *     title="Category Request",
 *     description="Category request body data",
 *     required={"name"},
 *     @OA\Property(property="name", type="string", example="Food"),
 *     @OA\Property(property="description", type="string", example="Events related to food and dining.")
 * )
 */