# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events API",
        "description": "API endpoints for managing events",
        "version": "1.0.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events with optional filtering",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "organizer",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Event"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "operationId": "e958e1c6c275d5f9b7ebcd837903d0ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get a single event by ID",
                "operationId": "c0f602d773798994c228d99dbd63b842",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "operationId": "8b1dd7f57d8fbd2e85af2762950e0b1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "operationId": "deca25daa2d423da0ff6b6677d92545b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "All categories"
                    }
                }
            }
        },
        "/categories/create": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show form to create a new category",
                "operationId": "3ef7e0bc5399e20ab25ef92f99cb0dae",
                "responses": {
                    "200": {
                        "description": "Create category form"
                    }
                }
            }
        },
        "/categories": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    }
                }
            }
        },
        "/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show a category",
                "operationId": "1d1c5342ee2901551673abd31b49dc85",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "230b7cb656677a24caa0012ba818db7f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update an existing category",
                "operationId": "01d3c61b2eea33a27b0cd17d6a7abaae",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    }
                }
            }
        },
        "/categories/{category}/edit": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show form to edit a category",
                "operationId": "40305ea3ad8a499a88287e1188719cda",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/category"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edit category form"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EventCreateRequest": {
                "title": "Event Request",
                "description": "Event request body data",
                "required": [
                    "title",
                    "start_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Taste of Cultures"
                    },
                    "description": {
                        "type": "string",
                        "example": "International food and culture festival celebrating global diversity in Brussels."
                    },
                    "location": {
                        "type": "string",
                        "example": "Kaaistudio's, Brussels"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-15T09:00:00"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-17T18:00:00"
                    },
                    "organizer": {
                        "type": "string",
                        "example": "CultuurConnect"
                    }
                },
                "type": "object"
            },
            "CategoryRequest": {
                "title": "Category Request",
                "description": "Category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Food"
                    },
                    "description": {
                        "type": "string",
                        "example": "Events related to food and dining."
                    }
                },
                "type": "object"
            },
            "EventRequest": {
                "title": "Event Request",
                "description": "Event request body data",
                "required": [
                    "title",
                    "start_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Taste of Cultures"
                    },
                    "description": {
                        "type": "string",
                        "example": "International food and culture festival celebrating global diversity in Brussels."
                    },
                    "location": {
                        "type": "string",
                        "example": "Kaaistudio’s, Brussels"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-15T09:00:00"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-17T18:00:00"
                    },
                    "organizer": {
                        "type": "string",
                        "example": "CultuurConnect"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category Model",
                "description": "Category model data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Music"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Event": {
                "title": "Event Model",
                "description": "Event model data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Taste of Cultures"
                    },
                    "description": {
                        "type": "string",
                        "example": "International food and culture festival celebrating global diversity in Brussels."
                    },
                    "location": {
                        "type": "string",
                        "example": "Kaaistudio's, Brussels"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-15T09:00:00"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-17T18:00:00"
                    },
                    "organizer": {
                        "type": "string",
                        "example": "CultuurConnect"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "category": {
                "name": "category",
                "in": "path",
                "description": "ID of the category",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "tags": [
        {
            "name": "Events",
            "description": "API Endpoints for managing events"
        },
        {
            "name": "Categories",
            "description": "Categories"
        }
    ]
}